load_image_file() support vstack option. Fixes in the predictor.
Extended the final_visualizations.ipynb so it predicts first with the latest weights and then displays.

+++ data_saver.py
@@ -10,7 +10,7 @@ def save_image(image, file_path):
 def save_image_from_tensor(image, file_path, multiply=False):
     multiplier = 1
     if multiply:
-        multiplier = 255
+        multiplier = np.uint8(255)
 
     result = Image.fromarray((image * multiplier).astype(np.uint8))
     result.save(file_path)
